third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-color-mix-function-expected.txt

This is a testharness.js-based test.
Found 474 tests; 289 PASS, 185 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), hsl(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 84 +/- 0.0001, expected 84 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 46
FAIL Property color value 'color-mix(in hsl, 25% hsl(120deg 10% 20%), hsl(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), 25% hsl(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 61 +/- 0.0001, expected 61 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), hsl(30deg 30% 40%) 25%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 61 +/- 0.0001, expected 61 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40%) 75%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 30%, hsl(30deg 30% 40%) 90%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 12.5%, hsl(30deg 30% 40%) 37.5%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 0%, hsl(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 133 +/- 0.0001, expected 133 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 95 +/- 0.0001, expected 95 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
FAIL Property color value 'color-mix(in hsl, 25% hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 121 +/- 0.0001, expected 121 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), 25% hsl(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 68 +/- 0.0001, expected 68 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8) 25%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 68 +/- 0.0001, expected 68 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 25%, hsl(30deg 30% 40% / .8) 75%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 121 +/- 0.0001, expected 121 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 30%, hsl(30deg 30% 40% / .8) 90%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 121 +/- 0.0001, expected 121 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 12.5%, hsl(30deg 30% 40% / .8) 37.5%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 121 +/- 0.0001, expected 121 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 0%, hsl(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 133 +/- 0.0001, expected 133 but got 46
FAIL Property color value 'color-mix(in hsl, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 149
FAIL Property color value 'color-mix(in hsl, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 191
FAIL Property color value 'color-mix(in hsl, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 170
FAIL Property color value 'color-mix(in hsl, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 64
FAIL Property color value 'color-mix(in hsl, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 64 +/- 0.0001, expected 64 but got 106
FAIL Property color value 'color-mix(in hsl, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 85 +/- 0.0001, expected 85 but got 149
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 149
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 191
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 170
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 64
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 64 +/- 0.0001, expected 64 but got 106
FAIL Property color value 'color-mix(in hsl shorter hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 85 +/- 0.0001, expected 85 but got 149
FAIL Property color value 'color-mix(in hsl longer hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl longer hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl longer hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl longer hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl longer hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl longer hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 149
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 64
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl increasing hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 85 +/- 0.0001, expected 85 but got 149
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 170 +/- 0.0001, expected 170 but got 191
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 85 +/- 0.0001, expected 85 but got 170
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 64 +/- 0.0001, expected 64 but got 106
FAIL Property color value 'color-mix(in hsl decreasing hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 64 +/- 0.0001, expected 64 but got 191
FAIL Property color value 'color-mix(in hsl specified hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hsl specified hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hsl specified hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hsl specified hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hsl specified hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hsl specified hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' assert_true: 'color-mix(in hsl specified hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))' is a supported value for color. expected true got false
PASS Property color value 'color-mix(in hsl, hsl(none none none), hsl(none none none))'
FAIL Property color value 'color-mix(in hsl, hsl(none none none), hsl(30deg 40% 80%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 224 +/- 0.0001, expected 224 but got 0
PASS Property color value 'color-mix(in hsl, hsl(120deg 20% 40%), hsl(none none none))'
FAIL Property color value 'color-mix(in hsl, hsl(120deg 20% none), hsl(30deg 40% 60%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 168 +/- 0.0001, expected 168 but got 0
FAIL Property color value 'color-mix(in hsl, hsl(120deg 20% 40%), hsl(30deg 20% none))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 112 +/- 0.0001, expected 112 but got 82
FAIL Property color value 'color-mix(in hsl, hsl(none 20% 40%), hsl(30deg none 80%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 173 +/- 0.0001, expected 173 but got 122
FAIL Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
FAIL Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40% / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 143 +/- 0.0001, expected 143 but got 61
FAIL Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40% / none))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 143 +/- 0.0001, expected 143 but got 61
FAIL Property color value 'color-mix(in hsl, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 249 +/- 0.0001, expected 249 but got 1
FAIL Property color value 'color-mix(in hsl, lab(100 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hsl, lab(0 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 42 +/- 0.0001, expected 42 but got 0
FAIL Property color value 'color-mix(in hsl, lch(100 116 334) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hsl, lch(0 116 334) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 42 +/- 0.0001, expected 42 but got 0
FAIL Property color value 'color-mix(in hsl, oklab(100 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hsl, oklab(0 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0 +/- 0.0001, expected 0 but got 0.365
FAIL Property color value 'color-mix(in hsl, oklch(100 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hsl, oklch(0 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0 +/- 0.0001, expected 0 but got 0.399
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 147 +/- 0.0001, expected 147 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 25%, hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, 25% hwb(120deg 10% 20%), hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), 25% hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 96 +/- 0.0001, expected 96 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% 40%) 25%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 96 +/- 0.0001, expected 96 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 25%, hwb(30deg 30% 40%) 75%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 30%, hwb(30deg 30% 40%) 90%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 12.5%, hwb(30deg 30% 40%) 37.5%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 0%, hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 153 +/- 0.0001, expected 153 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 143 +/- 0.0001, expected 143 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 25%, hwb(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 160 +/- 0.0001, expected 160 but got 26
FAIL Property color value 'color-mix(in hwb, 25% hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 160 +/- 0.0001, expected 160 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), 25% hwb(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8) 25%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 98 +/- 0.0001, expected 98 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 25%, hwb(30deg 30% 40% / .8) 75%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 160 +/- 0.0001, expected 160 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 30%, hwb(30deg 30% 40% / .8) 90%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 160 +/- 0.0001, expected 160 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 12.5%, hwb(30deg 30% 40% / .8) 37.5%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 160 +/- 0.0001, expected 160 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 0%, hwb(30deg 30% 40% / .8))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 153 +/- 0.0001, expected 153 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 128
FAIL Property color value 'color-mix(in hwb, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 153
FAIL Property color value 'color-mix(in hwb, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 140
FAIL Property color value 'color-mix(in hwb, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 77
FAIL Property color value 'color-mix(in hwb, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 77 +/- 0.0001, expected 77 but got 102
FAIL Property color value 'color-mix(in hwb, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 89 +/- 0.0001, expected 89 but got 127
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 128
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 153
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 140
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 77
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 77 +/- 0.0001, expected 77 but got 102
FAIL Property color value 'color-mix(in hwb shorter hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 89 +/- 0.0001, expected 89 but got 127
FAIL Property color value 'color-mix(in hwb longer hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb longer hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb longer hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb longer hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb longer hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb longer hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 128
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 77
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb increasing hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 89 +/- 0.0001, expected 89 but got 127
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 140 +/- 0.0001, expected 140 but got 153
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 89 +/- 0.0001, expected 89 but got 140
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 77 +/- 0.0001, expected 77 but got 102
FAIL Property color value 'color-mix(in hwb decreasing hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 77 +/- 0.0001, expected 77 but got 153
FAIL Property color value 'color-mix(in hwb specified hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hwb specified hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hwb specified hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hwb specified hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hwb specified hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in hwb specified hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' assert_true: 'color-mix(in hwb specified hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))' is a supported value for color. expected true got false
PASS Property color value 'color-mix(in hwb, hwb(none none none), hwb(none none none))'
FAIL Property color value 'color-mix(in hwb, hwb(none none none), hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 153 +/- 0.0001, expected 153 but got 255
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(none none none))'
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% none), hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 128 +/- 0.0001, expected 128 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% none))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 166 +/- 0.0001, expected 166 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(none 10% 20%), hwb(30deg none 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 179 +/- 0.0001, expected 179 but got 204
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40%))' assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40% / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 147 +/- 0.0001, expected 147 but got 26
FAIL Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40% / none))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 147 +/- 0.0001, expected 147 but got 26
FAIL Property color value 'color-mix(in hwb, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 249 +/- 0.0001, expected 249 but got 1
FAIL Property color value 'color-mix(in hwb, lab(100 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hwb, lab(0 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 42 +/- 0.0001, expected 42 but got 0
FAIL Property color value 'color-mix(in hwb, lch(100 116 334) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hwb, lch(0 116 334) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 42 +/- 0.0001, expected 42 but got 0
FAIL Property color value 'color-mix(in hwb, oklab(100 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hwb, oklab(0 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0 +/- 0.0001, expected 0 but got 0.365
FAIL Property color value 'color-mix(in hwb, oklch(100 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 0.0001, expected 255 but got 100
FAIL Property color value 'color-mix(in hwb, oklch(0 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0 +/- 0.0001, expected 0 but got 0.399
PASS Property color value 'color-mix(in lch, lch(10 20 30deg), lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg) 25%, lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, 25% lch(10 20 30deg), lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg), 25% lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg), lch(50 60 70deg) 25%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg) 25%, lch(50 60 70deg) 75%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg) 30%, lch(50 60 70deg) 90%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg) 12.5%, lch(50 60 70deg) 37.5%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg) 0%, lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4), lch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4) 25%, lch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in lch, 25% lch(10 20 30deg / .4), lch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4), 25% lch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4), lch(50 60 70deg / .8) 25%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4) 25%, lch(50 60 70deg / .8) 75%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4) 30%, lch(50 60 70deg / .8) 90%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4) 12.5%, lch(50 60 70deg / .8) 37.5%)'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / .4) 0%, lch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in lch, lch(100 0 40deg), lch(100 0 60deg))'
PASS Property color value 'color-mix(in lch, lch(100 0 60deg), lch(100 0 40deg))'
PASS Property color value 'color-mix(in lch, lch(100 0 50deg), lch(100 0 330deg))'
PASS Property color value 'color-mix(in lch, lch(100 0 330deg), lch(100 0 50deg))'
PASS Property color value 'color-mix(in lch, lch(100 0 20deg), lch(100 0 320deg))'
PASS Property color value 'color-mix(in lch, lch(100 0 320deg), lch(100 0 20deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 40deg), lch(100 0 60deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 60deg), lch(100 0 40deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 50deg), lch(100 0 330deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 330deg), lch(100 0 50deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 20deg), lch(100 0 320deg))'
PASS Property color value 'color-mix(in lch shorter hue, lch(100 0 320deg), lch(100 0 20deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 40deg), lch(100 0 60deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 60deg), lch(100 0 40deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 50deg), lch(100 0 330deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 330deg), lch(100 0 50deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 20deg), lch(100 0 320deg))'
PASS Property color value 'color-mix(in lch longer hue, lch(100 0 320deg), lch(100 0 20deg))'
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 40deg), lch(100 0 60deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 50 +/- 0.0001, expected 50 but got 230
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 60deg), lch(100 0 40deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 230 +/- 0.0001, expected 230 but got 50
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 50deg), lch(100 0 330deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 190 +/- 0.0001, expected 190 but got 10
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 330deg), lch(100 0 50deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 10 +/- 0.0001, expected 10 but got 190
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 20deg), lch(100 0 320deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 170 +/- 0.0001, expected 170 but got 350
FAIL Property color value 'color-mix(in lch increasing hue, lch(100 0 320deg), lch(100 0 20deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 350 +/- 0.0001, expected 350 but got 170
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 40deg), lch(100 0 60deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 230 +/- 0.0001, expected 230 but got 50
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 60deg), lch(100 0 40deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 50 +/- 0.0001, expected 50 but got 230
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 50deg), lch(100 0 330deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 10 +/- 0.0001, expected 10 but got 190
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 330deg), lch(100 0 50deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 190 +/- 0.0001, expected 190 but got 10
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 20deg), lch(100 0 320deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 350 +/- 0.0001, expected 350 but got 170
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100 0 320deg), lch(100 0 20deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 170 +/- 0.0001, expected 170 but got 350
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 40deg), lch(100 0 60deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 40deg), lch(100 0 60deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 60deg), lch(100 0 40deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 60deg), lch(100 0 40deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 50deg), lch(100 0 330deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 50deg), lch(100 0 330deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 330deg), lch(100 0 50deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 330deg), lch(100 0 50deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 20deg), lch(100 0 320deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 20deg), lch(100 0 320deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in lch specified hue, lch(100 0 320deg), lch(100 0 20deg))' assert_true: 'color-mix(in lch specified hue, lch(100 0 320deg), lch(100 0 20deg))' is a supported value for color. expected true got false
PASS Property color value 'color-mix(in lch, lch(none none none), lch(none none none))'
PASS Property color value 'color-mix(in lch, lch(none none none), lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg), lch(none none none))'
PASS Property color value 'color-mix(in lch, lch(10 20 none), lch(50 60 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg), lch(50 60 none))'
PASS Property color value 'color-mix(in lch, lch(none 20 30deg), lch(50 none 70deg))'
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / none), lch(50 60 70deg))'
FAIL Property color value 'color-mix(in lch, lch(10 20 30deg / none), lch(50 60 70deg / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 30 +/- 0.0001, expected 30 but got 23.3333
PASS Property color value 'color-mix(in lch, lch(10 20 30deg / none), lch(50 60 70deg / none))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg), oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg) 25%, oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, 25% oklch(10 20 30deg), oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg), 25% oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg), oklch(50 60 70deg) 25%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg) 25%, oklch(50 60 70deg) 75%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg) 30%, oklch(50 60 70deg) 90%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg) 12.5%, oklch(50 60 70deg) 37.5%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg) 0%, oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4), oklch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4) 25%, oklch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in oklch, 25% oklch(10 20 30deg / .4), oklch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4), 25% oklch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4), oklch(50 60 70deg / .8) 25%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4) 25%, oklch(50 60 70deg / .8) 75%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4) 30%, oklch(50 60 70deg / .8) 90%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4) 12.5%, oklch(50 60 70deg / .8) 37.5%)'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / .4) 0%, oklch(50 60 70deg / .8))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 40deg), oklch(100 0 60deg))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 60deg), oklch(100 0 40deg))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 50deg), oklch(100 0 330deg))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 330deg), oklch(100 0 50deg))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 20deg), oklch(100 0 320deg))'
PASS Property color value 'color-mix(in oklch, oklch(100 0 320deg), oklch(100 0 20deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 40deg), oklch(100 0 60deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 60deg), oklch(100 0 40deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 50deg), oklch(100 0 330deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 330deg), oklch(100 0 50deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 20deg), oklch(100 0 320deg))'
PASS Property color value 'color-mix(in oklch shorter hue, oklch(100 0 320deg), oklch(100 0 20deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 40deg), oklch(100 0 60deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 60deg), oklch(100 0 40deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 50deg), oklch(100 0 330deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 330deg), oklch(100 0 50deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 20deg), oklch(100 0 320deg))'
PASS Property color value 'color-mix(in oklch longer hue, oklch(100 0 320deg), oklch(100 0 20deg))'
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 40deg), oklch(100 0 60deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 50 +/- 0.0001, expected 50 but got 230
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 60deg), oklch(100 0 40deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 230 +/- 0.0001, expected 230 but got 50
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 50deg), oklch(100 0 330deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 190 +/- 0.0001, expected 190 but got 10
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 330deg), oklch(100 0 50deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 10 +/- 0.0001, expected 10 but got 190
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 20deg), oklch(100 0 320deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 170 +/- 0.0001, expected 170 but got 350
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100 0 320deg), oklch(100 0 20deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 350 +/- 0.0001, expected 350 but got 170
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 40deg), oklch(100 0 60deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 230 +/- 0.0001, expected 230 but got 50
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 60deg), oklch(100 0 40deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 50 +/- 0.0001, expected 50 but got 230
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 50deg), oklch(100 0 330deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 10 +/- 0.0001, expected 10 but got 190
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 330deg), oklch(100 0 50deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 190 +/- 0.0001, expected 190 but got 10
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 20deg), oklch(100 0 320deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 350 +/- 0.0001, expected 350 but got 170
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100 0 320deg), oklch(100 0 20deg))' assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 170 +/- 0.0001, expected 170 but got 350
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 40deg), oklch(100 0 60deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 40deg), oklch(100 0 60deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 60deg), oklch(100 0 40deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 60deg), oklch(100 0 40deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 50deg), oklch(100 0 330deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 50deg), oklch(100 0 330deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 330deg), oklch(100 0 50deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 330deg), oklch(100 0 50deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 20deg), oklch(100 0 320deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 20deg), oklch(100 0 320deg))' is a supported value for color. expected true got false
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100 0 320deg), oklch(100 0 20deg))' assert_true: 'color-mix(in oklch specified hue, oklch(100 0 320deg), oklch(100 0 20deg))' is a supported value for color. expected true got false
PASS Property color value 'color-mix(in oklch, oklch(none none none), oklch(none none none))'
PASS Property color value 'color-mix(in oklch, oklch(none none none), oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg), oklch(none none none))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 none), oklch(50 60 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg), oklch(50 60 none))'
PASS Property color value 'color-mix(in oklch, oklch(none 20 30deg), oklch(50 none 70deg))'
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / none), oklch(50 60 70deg))'
FAIL Property color value 'color-mix(in oklch, oklch(10 20 30deg / none), oklch(50 60 70deg / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 30 +/- 0.0001, expected 30 but got 23.3333
PASS Property color value 'color-mix(in oklch, oklch(10 20 30deg / none), oklch(50 60 70deg / none))'
PASS Property color value 'color-mix(in lab, lab(10 20 30), lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30) 25%, lab(50 60 70))'
PASS Property color value 'color-mix(in lab, 25% lab(10 20 30), lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30), 25% lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30), lab(50 60 70) 25%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30) 25%, lab(50 60 70) 75%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30) 30%, lab(50 60 70) 90%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30) 12.5%, lab(50 60 70) 37.5%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30) 0%, lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4), lab(50 60 70 / .8))'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4) 25%, lab(50 60 70 / .8))'
PASS Property color value 'color-mix(in lab, 25% lab(10 20 30 / .4), lab(50 60 70 / .8))'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4), 25% lab(50 60 70 / .8))'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4), lab(50 60 70 / .8) 25%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4) 25%, lab(50 60 70 / .8) 75%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4) 30%, lab(50 60 70 / .8) 90%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4) 12.5%, lab(50 60 70 / .8) 37.5%)'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / .4) 0%, lab(50 60 70 / .8))'
PASS Property color value 'color-mix(in lab, lab(none none none), lab(none none none))'
PASS Property color value 'color-mix(in lab, lab(none none none), lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30), lab(none none none))'
PASS Property color value 'color-mix(in lab, lab(10 20 none), lab(50 60 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30), lab(50 60 none))'
PASS Property color value 'color-mix(in lab, lab(none 20 30), lab(50 none 70))'
PASS Property color value 'color-mix(in lab, lab(10 20 30 / none), lab(50 60 70))'
FAIL Property color value 'color-mix(in lab, lab(10 20 30 / none), lab(50 60 70 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 30 +/- 0.0001, expected 30 but got 23.3333
PASS Property color value 'color-mix(in lab, lab(10 20 30 / none), lab(50 60 70 / none))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30), oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30) 25%, oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, 25% oklab(10 20 30), oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30), 25% oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30), oklab(50 60 70) 25%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30) 25%, oklab(50 60 70) 75%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30) 30%, oklab(50 60 70) 90%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30) 12.5%, oklab(50 60 70) 37.5%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30) 0%, oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4), oklab(50 60 70 / .8))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4) 25%, oklab(50 60 70 / .8))'
PASS Property color value 'color-mix(in oklab, 25% oklab(10 20 30 / .4), oklab(50 60 70 / .8))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4), 25% oklab(50 60 70 / .8))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4), oklab(50 60 70 / .8) 25%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4) 25%, oklab(50 60 70 / .8) 75%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4) 30%, oklab(50 60 70 / .8) 90%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4) 12.5%, oklab(50 60 70 / .8) 37.5%)'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / .4) 0%, oklab(50 60 70 / .8))'
PASS Property color value 'color-mix(in oklab, oklab(none none none), oklab(none none none))'
PASS Property color value 'color-mix(in oklab, oklab(none none none), oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30), oklab(none none none))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 none), oklab(50 60 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30), oklab(50 60 none))'
PASS Property color value 'color-mix(in oklab, oklab(none 20 30), oklab(50 none 70))'
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / none), oklab(50 60 70))'
FAIL Property color value 'color-mix(in oklab, oklab(10 20 30 / none), oklab(50 60 70 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 30 +/- 0.0001, expected 30 but got 23.3333
PASS Property color value 'color-mix(in oklab, oklab(10 20 30 / none), oklab(50 60 70 / none))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 25%, color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, 25% color(srgb .1 .2 .3), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), 25% color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 25%, color(srgb .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 30%, color(srgb .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 12.5%, color(srgb .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 0%, color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .5), color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 25%, color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, 25% color(srgb .1 .2 .3 / .4), color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4), color(srgb .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4), 25% color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 25%, color(srgb .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 30%, color(srgb .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 12.5%, color(srgb .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 0%, color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb 2 3 4 / 5), color(srgb 4 6 8 / 10))'
PASS Property color value 'color-mix(in srgb, color(srgb -2 -3 -4), color(srgb -4 -6 -8))'
PASS Property color value 'color-mix(in srgb, color(srgb -2 -3 -4 / -5), color(srgb -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in srgb, color(srgb none none none), color(srgb none none none))'
PASS Property color value 'color-mix(in srgb, color(srgb none none none), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb none none none))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 none), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 none))'
PASS Property color value 'color-mix(in srgb, color(srgb none .2 .3), color(srgb .5 none .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7))'
FAIL Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.3 +/- 0.0001, expected 0.3 but got 0.233333
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7 / none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 25%, color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, 25% color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), 25% color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 25%, color(srgb-linear .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 30%, color(srgb-linear .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 12.5%, color(srgb-linear .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 0%, color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .5), color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 25%, color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, 25% color(srgb-linear .1 .2 .3 / .4), color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4), color(srgb-linear .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4), 25% color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 25%, color(srgb-linear .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 30%, color(srgb-linear .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 12.5%, color(srgb-linear .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 0%, color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear 2 3 4 / 5), color(srgb-linear 4 6 8 / 10))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear -2 -3 -4), color(srgb-linear -4 -6 -8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear -2 -3 -4 / -5), color(srgb-linear -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none none none), color(srgb-linear none none none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none none none), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear none none none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 none), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none .2 .3), color(srgb-linear .5 none .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7))'
FAIL Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.3 +/- 0.0001, expected 0.3 but got 0.233333
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 25%, color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, 25% color(xyz .1 .2 .3), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), 25% color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 25%, color(xyz .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 30%, color(xyz .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 12.5%, color(xyz .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 0%, color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .5), color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 25%, color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, 25% color(xyz .1 .2 .3 / .4), color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4), color(xyz .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4), 25% color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 25%, color(xyz .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 30%, color(xyz .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 12.5%, color(xyz .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 0%, color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz 2 3 4 / 5), color(xyz 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz, color(xyz -2 -3 -4), color(xyz -4 -6 -8))'
PASS Property color value 'color-mix(in xyz, color(xyz -2 -3 -4 / -5), color(xyz -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz, color(xyz none none none), color(xyz none none none))'
PASS Property color value 'color-mix(in xyz, color(xyz none none none), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz none none none))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 none), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 none))'
PASS Property color value 'color-mix(in xyz, color(xyz none .2 .3), color(xyz .5 none .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7))'
FAIL Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.3 +/- 0.0001, expected 0.3 but got 0.233333
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 25%, color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, 25% color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), 25% color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 25%, color(xyz-d50 .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 30%, color(xyz-d50 .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 12.5%, color(xyz-d50 .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 0%, color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .5), color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 25%, color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, 25% color(xyz-d50 .1 .2 .3 / .4), color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4), color(xyz-d50 .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4), 25% color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 25%, color(xyz-d50 .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 30%, color(xyz-d50 .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 12.5%, color(xyz-d50 .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 0%, color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 2 3 4 / 5), color(xyz-d50 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 -2 -3 -4), color(xyz-d50 -4 -6 -8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 -2 -3 -4 / -5), color(xyz-d50 -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none none none), color(xyz-d50 none none none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none none none), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 none none none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 none), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none .2 .3), color(xyz-d50 .5 none .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7))'
FAIL Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.3 +/- 0.0001, expected 0.3 but got 0.233333
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 25%, color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, 25% color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), 25% color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 25%, color(xyz-d65 .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 30%, color(xyz-d65 .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 12.5%, color(xyz-d65 .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 0%, color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .5), color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 25%, color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, 25% color(xyz-d65 .1 .2 .3 / .4), color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4), color(xyz-d65 .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4), 25% color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 25%, color(xyz-d65 .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 30%, color(xyz-d65 .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 12.5%, color(xyz-d65 .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 0%, color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 2 3 4 / 5), color(xyz-d65 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 -2 -3 -4), color(xyz-d65 -4 -6 -8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 -2 -3 -4 / -5), color(xyz-d65 -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none none none), color(xyz-d65 none none none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none none none), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 none none none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 none), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none .2 .3), color(xyz-d65 .5 none .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7))'
FAIL Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7 / 0.5))' assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.3 +/- 0.0001, expected 0.3 but got 0.233333
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7 / none))'
Harness: the test ran to completion.
